+++ sentry-android-core/src/test/java/io/sentry/android/core/SentryAndroidOptionsTest.kt
@@ -3,6 +3,7 @@ package io.sentry.android.core
 import io.sentry.protocol.DebugImage
 import kotlin.test.Test
 import kotlin.test.assertEquals
+import kotlin.test.assertFalse
 import kotlin.test.assertNotNull
 import kotlin.test.assertTrue
 
@@ -60,6 +61,13 @@ class SentryAndroidOptionsTest {
         val sentryOptions = SentryAndroidOptions()
 
         assertTrue(sentryOptions.isEnableScopeSync)
+    }
+
+    @Test
+    fun `attach screenshots disabled by default for Android`() {
+        val sentryOptions = SentryAndroidOptions()
+
+        assertFalse(sentryOptions.isAttachScreenshot)
     }
 
     private class CustomDebugImagesLoader : IDebugImagesLoader {